Admin Update Bug fix

+++ DutyRoster1/src/dutyroster/ExportFile.java
@@ -35,6 +35,9 @@ public class ExportFile {
         
         file = fileChooser.showSaveDialog(mainStage);
         
+        if(file==null)
+            return;
+        
         boolean result = Files.deleteIfExists(file.toPath());
     
     }
